tests/testutils/site.py: Move the pip_sample_plugins() check to a common site.py
As we will be using the sample plugins package to carry some plugins we want
to use for general purpose testing of the API surface throughout buildstream,
we'll need to have this in a shared location.
